+++ src/components/editors/SceneEditor.tsx
@@ -52,6 +52,7 @@ import { BackgroundSymbolsEditor } from "components/forms/symbols/BackgroundSymb
 import { SymbolEditorWrapper } from "components/forms/symbols/SymbolEditorWrapper";
 import { ScriptEditorContext } from "components/script/ScriptEditorContext";
 import Alert, { AlertItem } from "components/library/Alert";
+import { sceneName } from "store/features/entities/entitiesHelpers";
 
 interface SceneEditorProps {
   id: string;
@@ -121,9 +122,6 @@ const getScriptKey = (
   }
   return "script";
 };
-
-const sceneName = (scene: Scene, sceneIndex: number) =>
-  scene.name ? scene.name : `Scene ${sceneIndex + 1}`;
 
 export const SceneEditor = ({ id, multiColumn }: SceneEditorProps) => {
   const scene = useSelector((state: RootState) =>