+++ src/components/editors/TriggerEditor.tsx
@@ -33,6 +33,7 @@ import { ClipboardTypeTriggers } from "store/features/clipboard/clipboardTypes";
 import { TriggerSymbolsEditor } from "components/forms/symbols/TriggerSymbolsEditor";
 import { SymbolEditorWrapper } from "components/forms/symbols/SymbolEditorWrapper";
 import { ScriptEditorContext } from "components/script/ScriptEditorContext";
+import { triggerName } from "store/features/entities/entitiesHelpers";
 
 interface TriggerEditorProps {
   id: string;
@@ -70,9 +71,6 @@ const getScriptKey = (tab: keyof typeof scriptTabs): TriggerScriptKey => {
   }
   return "script";
 };
-
-const triggerName = (trigger: Trigger, triggerIndex: number) =>
-  trigger.name ? trigger.name : `Trigger ${triggerIndex + 1}`;
 
 const tabs = Object.keys(scriptTabs);