+++ src/components/forms/SceneSelect.tsx
@@ -16,6 +16,7 @@ import { Scene } from "store/features/entities/entitiesTypes";
 import { RootState } from "store/configureStore";
 import styled from "styled-components";
 import editorActions from "store/features/editor/editorActions";
+import { sceneName } from "store/features/entities/entitiesHelpers";
 
 interface SceneSelectProps extends SelectCommonProps {
   name: string;
@@ -48,7 +49,7 @@ const sortByLabel = (a: SceneOption, b: SceneOption) => {
 
 const sceneToSceneOption = (scene: Scene, sceneIndex: number): SceneOption => ({
   value: scene.id,
-  label: scene.name ? scene.name : `Scene ${sceneIndex + 1}`,
+  label: sceneName(scene, sceneIndex),
   scene,
 });