+++ src/components/world/Scene.js
@@ -36,6 +36,7 @@ import entitiesActions from "store/features/entities/entitiesActions";
 import ScenePriorityMap from "./ScenePriorityMap";
 import SceneSlopePreview from "./SceneSlopePreview";
 import { SceneEventHelper } from "./SceneEventHelper";
+import { sceneName } from "store/features/entities/entitiesHelpers";
 
 const TILE_SIZE = 8;
 
@@ -135,7 +136,7 @@ class Scene extends Component {
       id,
       scene,
       visible,
-      sceneName,
+      name,
       image,
       width,
       height,
@@ -205,7 +206,7 @@ class Scene extends Component {
               "Scene__Label--Gray": labelColor === "gray",
             })}
           >
-            {sceneName}
+            {name}
           </div>
         </div>
         <div
@@ -377,7 +378,7 @@ Scene.propTypes = {
   selectScene: PropTypes.func.isRequired,
   moveSelectedEntity: PropTypes.func.isRequired,
   sceneHover: PropTypes.func.isRequired,
-  sceneName: PropTypes.string.isRequired,
+  name: PropTypes.string.isRequired,
   sceneFiltered: PropTypes.bool.isRequired,
   labelOffsetLeft: PropTypes.number.isRequired,
   labelOffsetRight: PropTypes.number.isRequired,
@@ -445,10 +446,10 @@ function mapStateToProps(state, props) {
     : 0;
 
   const searchTerm = state.editor.searchTerm;
-  const sceneName = scene.name || `Scene ${props.index + 1}`;
+  const name = sceneName(scene, props.index);
   const sceneFiltered =
     (searchTerm &&
-      sceneName.toUpperCase().indexOf(searchTerm.toUpperCase()) === -1 &&
+      name.toUpperCase().indexOf(searchTerm.toUpperCase()) === -1 &&
       scene.id !== searchTerm) ||
     false;
 
@@ -533,7 +534,7 @@ function mapStateToProps(state, props) {
     selected,
     dragging,
     hovered,
-    sceneName,
+    name,
     sceneFiltered,
     palettes,
     spritePalettes,